move handler ith handlers and initialize assStore needed by  api call
