+++ api/server/restapi/operations/add_asset.go
@@ -14,7 +14,6 @@ import (
 	"github.com/go-openapi/runtime/middleware"
 	"github.com/go-openapi/strfmt"
 	"github.com/go-openapi/swag"
-
 	"github.com/massalabs/station-massa-wallet/api/server/models"
 )
