+++ api/server/restapi/operations/add_asset_responses.go
@@ -9,7 +9,6 @@ import (
 	"net/http"
 
 	"github.com/go-openapi/runtime"
-
 	"github.com/massalabs/station-massa-wallet/api/server/models"
 )
