+++ internal/handler/api.go
@@ -7,12 +7,13 @@ import (
 	"github.com/massalabs/station-massa-wallet/api/server/restapi/operations"
 	"github.com/massalabs/station-massa-wallet/internal/handler/html"
 	walletHandler "github.com/massalabs/station-massa-wallet/internal/handler/wallet"
+	"github.com/massalabs/station-massa-wallet/pkg/assets"
 	"github.com/massalabs/station-massa-wallet/pkg/network"
 	"github.com/massalabs/station-massa-wallet/pkg/prompt"
 )
 
 // InitializeAPI initializes the API handlers
-func InitializeAPI(prompterApp prompt.WalletPrompterInterface, massaClient network.NodeFetcherInterface, gc gcache.Cache) (*operations.MassaWalletAPI, error) {
+func InitializeAPI(prompterApp prompt.WalletPrompterInterface, massaClient network.NodeFetcherInterface, assetsStore *assets.AssetsStore, gc gcache.Cache) (*operations.MassaWalletAPI, error) {
 	// Load the Swagger specification
 	swaggerSpec, err := loads.Analyzed(restapi.SwaggerJSON, "")
 	if err != nil {
@@ -26,7 +27,7 @@ func InitializeAPI(prompterApp prompt.WalletPrompterInterface, massaClient netwo
 	html.AppendEndpoints(api)
 
 	// Set wallet API endpoints
-	walletHandler.AppendEndpoints(api, prompterApp, massaClient, gc)
+	walletHandler.AppendEndpoints(api, prompterApp, massaClient, assetsStore, gc)
 
 	return api, nil
 }
